hud (ui)

+++ lwjgl/src/main/java/io/github/heathensoft/jlib/lwjgl/utils/Input.java
@@ -69,7 +69,7 @@ public class Input extends InputProcessor {
     }
 
     @Override
-    protected void on_mouse_scroll(int amount) {
+    protected void on_mouse_scroll(float amount) {
         mouse.on_mouse_scroll(amount);
     }
